+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [1.0.0] - 2021-01-27
+
 ### Added
 
 - Reducers
@@ -71,3 +73,4 @@ and this project adheres to
 [0.2.0]: https://github.com/jajaperson/iterable-utilities/releases/tag/v0.2.0
 [0.1.1]: https://github.com/jajaperson/iterable-utilities/releases/tag/v0.1.1
 [0.1.0]: https://github.com/jajaperson/iterable-utilities/releases/tag/v0.1.0
+[1.0.0]: https://github.com/jajaperson/iterable-utilities/releases/tag/v1.0.0